:bug: [#40] fix: 기존 backupfile 패키지를 file 패키지로 병합함. - 현재 사용하지 않는 S3 관련 로직들 삭제 - 추후 s3 관련 로직을 다시 작성할 예정
