+++ src/main/java/com/example/mody/domain/file/repository/BackupFileRepository.java
@@ -1,6 +1,6 @@
-package com.example.mody.domain.backupimage.repository;
+package com.example.mody.domain.file.repository;
 
-import com.example.mody.domain.backupimage.entity.BackupFile;
+import com.example.mody.domain.file.entity.BackupFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
